Updated UI to look better

+++ Assets/Scripts/player_inventory_values.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class player_inventory_values : MonoBehaviour {
 
-    public int MoneyReserve = 1000;
+    public float MoneyReserve = 1000;
     public int[] AmountOfItems = { 0, 0, 0, 0, 0 }; // First three indices are plants, fourth and fifth indices are fertilizers
 
 }
